refactor(registry): split MetadataStore into four aggregate stores
Registry RFC phase B. Row dataclasses move to rows.py; schema bootstrap
(lock, statements minus api_tokens, migrations, integer flags) runs once
in store_schema.open_stores; package/result/org/inbox stores carry their
own aggregate methods over the shared dialect adapter, each described by
a narrow protocol in protocols.py. store.py keeps a duck-typed
MetadataStore facade (open_stores + delegation) and re-exports so no
caller changes yet - phase C removes the god interface. The no-SQL-
literal architecture tests now scan every store*.py.
