+++ services/registry/clock.py
@@ -1,0 +1,9 @@
+"""Monotonic-enough wall clock for Registry timestamps (epoch seconds)."""
+
+from __future__ import annotations
+
+import time
+
+
+def now() -> float:
+    return time.time()
