+++ services/registry/rows.py
@@ -1,0 +1,197 @@
+"""Row dataclasses shared by the Registry aggregate stores."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+# ---------------------------------------------------------------------------
+# Rows
+# ---------------------------------------------------------------------------
+
+
+@dataclass(frozen=True, slots=True)
+class ReleaseRow:
+    dataset_id: str
+    version: str
+    visibility: str
+    package_digest: str
+    blob_digest: str
+    size: int
+    media_type: str
+    created_at: float
+    org_id: str | None = None
+    uploaded_by: str = ""
+
+
+@dataclass(frozen=True, slots=True)
+class AttemptResultRow:
+    """Sealed Attempt evidence bundle metadata (not a Dataset package)."""
+
+    run_id: str
+    dataset_id: str
+    task_id: str
+    lock_digest: str
+    status: str
+    visibility: str
+    blob_digest: str
+    size: int
+    created_at: float
+    dataset_version: str = ""
+    uploaded_by: str = ""
+    # Optional link to parent suite/job (#43); empty on standalone uploads.
+    suite_run_id: str = ""
+    environment: str = ""
+    agent_label: str = ""
+    model_label: str = ""
+    score: float | None = None
+
+
+@dataclass(frozen=True, slots=True)
+class SuiteResultRow:
+    """Suite/job result row: aggregates + per-task refs (not suite PASS).
+
+    Observational leaderboard input for Hub SPA (#22 S5). PASS remains
+    per-task evaluator only.
+    """
+
+    suite_run_id: str
+    dataset_id: str
+    dataset_version: str
+    visibility: str
+    pass_rate: float
+    mean_score: float
+    metrics_json: str
+    tasks_json: str
+    agent_label: str
+    model_label: str
+    blob_digest: str
+    size: int
+    exit_code: int
+    created_at: float
+    # #42 config comparability (optional; empty/default on legacy rows)
+    config_json: str = "{}"
+    uploaded_by: str = ""
+    complete: bool = False
+    bound_kind: str = "unknown"
+    task_set_digest: str = ""
+    board_listed: bool = False
+
+
+@dataclass(frozen=True, slots=True)
+class ResourceRequestRow:
+    request_id: str
+    kind: str
+    status: str
+    suite_run_id: str
+    dataset_id: str
+    applicant: str
+    owner_org_id: str
+    agent_ref: str
+    created_at: float
+    decided_at: float | None = None
+    decided_by: str = ""
+    canonical_model: str = ""
+
+
+@dataclass(frozen=True, slots=True)
+class DraftRow:
+    """One current draft slot per dataset (not a release)."""
+
+    dataset_id: str
+    org_id: str
+    visibility: str
+    package_digest: str
+    blob_digest: str
+    size: int
+    media_type: str
+    package_kind: str
+    uploaded_by: str
+    updated_at: float
+
+    def as_release(self) -> ReleaseRow:
+        return ReleaseRow(
+            dataset_id=self.dataset_id,
+            version="draft",
+            visibility=self.visibility,
+            package_digest=self.package_digest,
+            blob_digest=self.blob_digest,
+            size=self.size,
+            media_type=self.media_type,
+            created_at=self.updated_at,
+            org_id=self.org_id,
+            uploaded_by=self.uploaded_by,
+        )
+
+
+@dataclass(frozen=True, slots=True)
+class DatasetAclRow:
+    dataset_id: str
+    user_id: str
+    role: str
+    created_at: float
+
+
+@dataclass(frozen=True, slots=True)
+class OrgRow:
+    org_id: str
+    name: str
+    display_name: str
+    description: str
+    is_claimable: bool
+    created_at: float
+    icon_key: str = ""
+    icon_github: str = ""
+
+
+@dataclass(frozen=True, slots=True)
+class MembershipRow:
+    org_id: str
+    user_id: str
+    role: str
+    created_at: float
+
+
+@dataclass(frozen=True, slots=True)
+class UserProfileRow:
+    """GitHub profile snapshot written at Registry login (not a credential)."""
+
+    user_id: str
+    display_name: str
+    avatar_url: str
+    github_id: str
+    description: str
+    updated_at: float
+
+
+@dataclass(frozen=True, slots=True)
+class ResultShareRow:
+    result_kind: str  # attempt | suite
+    result_id: str
+    target_type: str  # org | user
+    target_id: str
+    created_at: float
+
+
+@dataclass(frozen=True, slots=True)
+class OrgInviteKeyRow:
+    """Org invite key — at rest: token_hash + token_prefix only (no plaintext).
+
+    Full secret is generated by the HTTP layer and returned **once** on create;
+    redeem looks up by SHA-256 hash. Owner list/revoke never re-materialize the
+    secret (industry default for invite/API keys).
+    """
+
+    key_id: str
+    org_id: str
+    token_hash: str
+    token_prefix: str
+    created_by: str
+    max_uses: int | None
+    use_count: int
+    expires_at: float | None
+    revoked_at: float | None
+    created_at: float
+
+
+# ---------------------------------------------------------------------------
+# Metadata (packages + attempt results + orgs + shares)
