+++ tests/registry/test_dialect_queries.py
@@ -142,17 +142,16 @@ def test_align_integer_flag_rejects_bad_ident() -> None:
 def test_store_has_no_sql_literals() -> None:
     from pathlib import Path
 
-    text = (Path(__file__).resolve().parents[2] / "services" / "registry" / "store.py").read_text(
-        encoding="utf-8"
-    )
     needles = ("DELETE FROM", "INSERT INTO", "CREATE TABLE", "UPDATE ")
+    registry = Path(__file__).resolve().parents[2] / "services" / "registry"
     offenders: list[str] = []
-    for i, line in enumerate(text.splitlines(), 1):
-        stripped = line.strip()
-        if stripped.startswith("#"):
-            continue
-        if any(n in line for n in needles):
-            offenders.append(f"{i}:{stripped}")
+    for path in sorted(registry.glob("store*.py")):
+        for i, line in enumerate(path.read_text(encoding="utf-8").splitlines(), 1):
+            stripped = line.strip()
+            if stripped.startswith("#"):
+                continue
+            if any(n in line for n in needles):
+                offenders.append(f"{path.name}:{i}:{stripped}")
     assert offenders == []
